requirements.txt
django
gunicorn
django-heroku
requests
pyjwt
